Night Music + Prune/Sprout SFX + Ambience
Night Music: Has an intro (when the timer ends) and then a loop. There is a discrete parameter for spirits (0-6) which adds an instrument with each addition. There is also a global continuous parameter for "Distance," which will make the music get muffled when you zoom in.

Prune and Leaves Sprout SFX: Three different versions, will be selected randomly each time it's triggered.

Night Ambience: Uses same Distance parameter as the music. When at a far distance, air noise is slightly louder, and when at a close distance, there are lots of little scuttles and rustles and foley that trigger at random locations and from scatterer instruments.